requirements.txt
requests
opencv-python
pytesseract
rapidfuzz
python-dotenv
BeautifulSoup4
flask
